feat: improve imguizmo, fix imgui enum issue

+++ dear-imgui/src/columns.rs
@@ -8,17 +8,17 @@ bitflags! {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
     pub struct OldColumnFlags: i32 {
         /// No flags
-        const NONE = sys::ImGuiOldColumnFlags_None;
+        const NONE = sys::ImGuiOldColumnFlags_None as i32;
         /// Disable column dividers
-        const NO_BORDER = sys::ImGuiOldColumnFlags_NoBorder;
+        const NO_BORDER = sys::ImGuiOldColumnFlags_NoBorder as i32;
         /// Disable resizing columns by dragging dividers
-        const NO_RESIZE = sys::ImGuiOldColumnFlags_NoResize;
+        const NO_RESIZE = sys::ImGuiOldColumnFlags_NoResize as i32;
         /// Disable column width preservation when the total width changes
-        const NO_PRESERVE_WIDTHS = sys::ImGuiOldColumnFlags_NoPreserveWidths;
+        const NO_PRESERVE_WIDTHS = sys::ImGuiOldColumnFlags_NoPreserveWidths as i32;
         /// Disable forcing columns to fit within window
-        const NO_FORCE_WITHIN_WINDOW = sys::ImGuiOldColumnFlags_NoForceWithinWindow;
+        const NO_FORCE_WITHIN_WINDOW = sys::ImGuiOldColumnFlags_NoForceWithinWindow as i32;
         /// Restore pre-1.51 behavior of extending the parent window contents size
-        const GROW_PARENT_CONTENTS_SIZE = sys::ImGuiOldColumnFlags_GrowParentContentsSize;
+        const GROW_PARENT_CONTENTS_SIZE = sys::ImGuiOldColumnFlags_GrowParentContentsSize as i32;
     }
 }
